requirements.txt
# Core
pydantic>=2.0
numpy>=1.24
pandas>=1.5
loguru>=0.7
tqdm>=4.65
psutil>=5.9
python-dotenv>=1.0
requests>=2.28

# ML / Embeddings
torch>=2.0
sentence-transformers>=2.2
faiss-cpu>=1.7           # use faiss-gpu for GPU support
scikit-learn>=1.2
scipy>=1.10
umap-learn>=0.5

# Graph
networkx>=3.0

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.14
dash>=2.10

# Web frameworks
flask>=2.3
fastapi>=0.100
uvicorn>=0.22

# Agent (MAG dataset)
neo4j>=5.0
torch-geometric>=2.3
openai>=1.0

# Optional -- uncomment as needed
# ray>=2.5              # parallel document processing
# cupy>=12.0            # GPU-accelerated UMAP
# cuml>=23.0            # GPU-accelerated UMAP
# cdlib>=0.3            # community detection
